Restructured GuiFunctions
