+++ MetaMorpheus/GUI/MetaDraw/MetaDraw.xaml.cs
@@ -1,6 +1,7 @@
 using Easy.Common.Extensions;
 using EngineLayer;
 using GuiFunctions;
+using GuiFunctions.MetaDraw.Chimeras;
 using MassSpectrometry;
 using Omics.Fragmentation;
 using OxyPlot;
