+++ MetaMorpheus/GuiFunctions/MetaDraw/MetaDrawSettings.cs
@@ -9,6 +9,7 @@ using System.IO;
 using System.Linq;
 using System.Windows.Media;
 using Readers;
+using GuiFunctions.MetaDraw.Chimeras;
 
 namespace GuiFunctions
 {
