+++ MetaMorpheus/Test/MetaDraw/ChimeraDrawnSequenceTests.cs
@@ -4,6 +4,7 @@ using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Shapes;
 using GuiFunctions;
+using GuiFunctions.MetaDraw.Chimeras;
 using NUnit.Framework;
 
 namespace Test.MetaDraw;
