+++ MetaMorpheus/Test/MetaDraw/ChimeraGroupViewModelTests.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using Chemistry;
 using EngineLayer;
 using GuiFunctions;
+using GuiFunctions.MetaDraw.Chimeras;
 using MassSpectrometry;
 using NUnit.Framework;
 using NUnit.Framework.Legacy;
