+++ MetaMorpheus/Test/MetaDraw/MetaDrawTest.cs
@@ -13,6 +13,7 @@ using Chemistry;
 using Easy.Common.Extensions;
 using EngineLayer;
 using GuiFunctions;
+using GuiFunctions.MetaDraw.Chimeras;
 using MassSpectrometry;
 using NUnit.Framework;
 using OxyPlot.Series;
